requirements.txt
scikit-learn
numpy
pandas
gradio
python-dotenv
requests
beautifulsoup4
whois
tldextract